feat: rps.js locale  적용
chore: 'totallynotagame'과 관련된 파일, 문자열을 'rps'로 변경

chore: index.html에 rps.css 추가
