+++ fe/index.html
@@ -7,7 +7,7 @@
     <title>SPA 예제</title>
     <link href="/node_modules/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet">
     <link href="/index.css" rel="stylesheet">
-    <link href="/totallynotagame.css" rel="stylesheet">
+    <link href="/rps.css" rel="stylesheet">
     <script type="module" src="/src/App.js" defer></script>
 </head>
