+++ fe/locale/en-US.json
@@ -6,10 +6,21 @@
 	"error": "Your game ran into a problem and needs to restart as soon as we're finished collecting some error info.",
 	"errorButton": "Got it, take me to the home screen",
 	"scoreboard": "scoreborad",
-	"totallynotagame": "totally not a game!",
+	"rps": "Rock, Paper, Scissors",
 	"lorem": "We the People of the United States, in Order to form a more perfect Union, establish Justice, insure domestic Tranquility, provide for the common defense, promote the general Welfare, and secure the Blessings of Liberty to ourselves and our Posterity, do ordain and establish this Constitution for the United States of America. All legislative Powers herein granted shall be vested in a Congress of the United States, which shall consist of a Senate and House of Representatives. The House of Representatives shall be composed of Members chosen every second Year by the People of the several States, and the Electors in each State shall have the Qualifications requisite for Electors of the most numerous Branch of the State Legislature.",
 	"result-t": "tournament record",
 	"result-d": "1:1 record",
 	"result-rps": "RPS record",
-	"loading": "loading..."
+	"loading": "loading...",
+
+    "rps-startMatching": "Start matching",
+    "rps-watingOpponent": "Waiting for opponent...",
+    "rps-cancelMatching": "Cancel matching",
+    "rps-gameMatched": "Game matched! Game starting...",
+    "rps-selection": "If you do not select in 10 seconds, a random selection will be made.",
+    "rps-watingResult": "Waiting for result...",
+    "rps-result-win": "You win!",
+    "rps-result-lose": "You lose!",
+    "rps-result-draw": "Draw!",
+    "ros-result-unknown": "Unknown result"
 }
