+++ fe/locale/ja-JP.json
@@ -6,10 +6,21 @@
 	"error":"ゲームに問題が発生したため、 再起動する必要があります。エラー情報を収集しています。 自動的に再起動します。",
 	"errorButton": "再起動",
 	"scoreboard": "スコアボード",
-	"totallynotagame": "決してゲームではありません！",
+	"rps": "じゃんけん",
 	"lorem": "日本国民は、正当に選挙された国会における代表者を通じて行動し、われらとわれらの子孫のために、諸国民との協和による成果と、わが国全土にわたつて自由のもたらす恵沢を確保し、政府の行為によつて再び戦争の惨禍が起ることのないやうにすることを決意し、ここに主権が国民に存することを宣言し、この憲法を確定する。そもそも国政は、国民の厳粛な信託によるものであつて、その権威は国民に由来し、その権力は国民の代表者がこれを行使し、その福利は国民がこれを享受する。これは人類普遍の原理であり、この憲法は、かかる原理に基くものである。われらは、これに反する一切の憲法、法令及び詔勅を排除する。日本国民は、恒久の平和を念願し、人間相互の関係を支配する崇高な理想を深く自覚するのであつて、平和を愛する諸国民の公正と信義に信頼して、われらの安全と生存を保持しようと決意した。われらは、平和を維持し、専制と隷従、圧迫と偏狭を地上から永遠に除去しようと努めてゐる国際社会において、名誉ある地位を占めたいと思ふ。われらは、全世界の国民が、ひとしく恐怖と欠乏から免かれ、平和のうちに生存する権利を有することを確認する。われらは、いづれの国家も、自国のことのみに専念して他国を無視してはならないのであつて、政治道徳の法則は、普遍的なものであり、この法則に従ふことは、自国の主権を維持し、他国と対等関係に立たうとする各国の責務であると信ずる。日本国民は、国家の名誉にかけ、全力をあげてこの崇高な理想と目的を達成することを誓ふ。",
 	"result-t": "トーナメント戦績",
 	"result-d": "1:1戦績",
 	"result-rps": "じゃんけん戦績",
-	"loading": "読み込み中..."
+	"loading": "読み込み中...",
+
+    "rps-startMatching": "マッチング開始",
+    "rps-watingOpponent": "対戦相手を待っています...",
+    "rps-cancelMatching": "マッチングキャンセル",
+    "rps-gameMatched": "ゲームがマッチしました！ゲームが開始されます...",
+    "rps-selection": "10秒以内に選択しない場合、ランダムに選択されます。",
+    "rps-watingResult": "結果を待っています...",
+    "rps-result-win": "あなたの勝ちです！",
+    "rps-result-lose": "あなたの負けです！",
+    "rps-result-draw": "引き分けです！",
+    "ros-result-unknown": "不明な結果"
 }
