+++ fe/locale/ko-KR.json
@@ -6,10 +6,21 @@
 	"error": "게임에 문제가 발생하여 다시 시작해야 합니다. 일부 오류 정보를 수집하고 있습니다. 그런 다음 자동으로 다시 시작합니다.",
 	"errorButton": "재시작",
 	"scoreboard": "스코어보드",
-	"totallynotagame": "절대로게임이아닙니다!",
+	"rps": "가위 바위 보",
 	"lorem": "유구한 역사와 전통에 빛나는 우리 대한국민은 3ㆍ1운동으로 건립된 대한민국임시정부의 법통과 불의에 항거한 4ㆍ19민주이념을 계승하고, 조국의 민주개혁과 평화적 통일의 사명에 입각하여 정의ㆍ인도와 동포애로써 민족의 단결을 공고히 하고, 모든 사회적 폐습과 불의를 타파하며, 자율과 조화를 바탕으로 자유민주적 기본질서를 더욱 확고히 하여 정치ㆍ경제ㆍ사회ㆍ문화의 모든 영역에 있어서 각인의 기회를 균등히 하고, 능력을 최고도로 발휘하게 하며, 자유와 권리에 따르는 책임과 의무를 완수하게 하여, 안으로는 국민생활의 균등한 향상을 기하고 밖으로는 항구적인 세계평화와 인류공영에 이바지함으로써 우리들과 우리들의 자손의 안전과 자유와 행복을 영원히 확보할 것을 다짐하면서 1948년 7월 12일에 제정되고 8차에 걸쳐 개정된 헌법을 이제 국회의 의결을 거쳐 국민투표에 의하여 개정한다.",
 	"result-t": "토너먼트 전적",
 	"result-d": "1:1 전적",
 	"result-rps": "가위바위보 전적",
-	"loading": "로딩중..."
+	"loading": "로딩중...",
+
+    "rps-startMatching": "매칭 시작",
+    "rps-watingOpponent": "상대방을 기다리는 중...",
+    "rps-cancelMatching": "매칭 취소",
+    "rps-gameMatched": "게임 매칭 완료! 게임 시작...",
+    "rps-selection": "10초 내에 선택하지 않으면 무작위 선택이 이루어집니다.",
+    "rps-watingResult": "결과를 기다리는 중...",
+    "rps-result-win": "당신이 이겼습니다!",
+    "rps-result-lose": "당신이 졌습니다!",
+    "rps-result-draw": "비겼습니다!",
+    "ros-result-unknown": "알 수 없는 결과"
 }
