+++ fe/src/App.js
@@ -45,7 +45,7 @@ const routes = {
         'dual': () => import('./pages/log/dual.js').then(module => module.render(app, navigate)),
         'rps': () => import('./pages/log/rps.js').then(module => module.render(app, navigate)),
     },
-    'totallynotagame': () => import('./pages/totallynotagame.js').then(module => module.render(app, navigate)),
+    'rps': () => import('./pages/rps.js').then(module => module.render(app, navigate)),
 };
 
 // Default error page
